[SO-181] refactor : DummyEntity global 폴더로 이동
