+++ src/test/java/swmaestro/spaceodyssey/weddingmate/domain/portfolio/controller/PortfolioControllerTest.java
@@ -26,11 +26,10 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import jakarta.persistence.EntityManager;
-import jakarta.transaction.Transactional;
 import swmaestro.spaceodyssey.weddingmate.domain.item.dto.ItemOrderDto;
 import swmaestro.spaceodyssey.weddingmate.domain.item.entity.Item;
 import swmaestro.spaceodyssey.weddingmate.domain.item.repository.ItemRepository;
-import swmaestro.spaceodyssey.weddingmate.domain.portfolio.PortfolioDummyEntity;
+import swmaestro.spaceodyssey.weddingmate.global.config.test.DummyEntity;
 import swmaestro.spaceodyssey.weddingmate.domain.portfolio.dto.PortfolioSaveReqDto;
 import swmaestro.spaceodyssey.weddingmate.domain.portfolio.dto.PortfolioUpdateReqDto;
 import swmaestro.spaceodyssey.weddingmate.domain.portfolio.entity.Portfolio;
@@ -43,7 +42,7 @@ import swmaestro.spaceodyssey.weddingmate.domain.users.repository.UsersRepositor
 @Sql("classpath:db/teardown.sql")
 @AutoConfigureMockMvc
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-public class PortfolioControllerTest extends PortfolioDummyEntity {
+public class PortfolioControllerTest extends DummyEntity {
 
 	@Autowired
 	private PortfolioService portfolioService;
